requirements.txt
flask==3.0.3
google-generativeai==0.5.4
pypdf2==3.0.1
python-docx==1.1.2
langchain==0.2.1
langchain-community==0.2.1
langchain-google-genai==1.0.5
chromadb==0.5.0
python-dotenv==1.0.1
flask-cors==4.0.0
humanize==4.9.0
Flask-JWT-Extended==4.6.0
gunicorn==22.0.0